Add files via upload
Welcome to my Code Lesley Portfolio!

This portfolio showcases my projects and achievements as a coding enthusiast and learner. As a part of the Code Lesley community, I'm committed to continuously improving my coding skills and exploring new technologies.

About Me

I'm Lesley Shahid Fuller, a Blockchain developer with a passion for coding and innovation. I'm excited to share my projects and experiences with you, and I'm always looking for opportunities to learn and grow.

Project 1: Voting Application

and There are certainly more to be added to my Profile